Tidy filter and issuer list

+++ src/SecuritiesService/Service/Securities/ByGroupService.php
@@ -41,10 +41,10 @@ class ByGroupService extends SecuritiesService
     public function sum(
         ParentGroup $group,
         SecuritiesFilter $filter = null
-    ): int {
+    ): float {
         $qb = $this->queryForScalar($group, $filter);
         $qb->select('sum(' . self::TBL . '.moneyRaised)');
-        return (int) $qb->getQuery()->getSingleScalarResult();
+        return (float) $qb->getQuery()->getSingleScalarResult();
     }
 
     public function issuanceYears(
